added float cast

+++ src/double_pendulum.cpp
@@ -29,10 +29,10 @@ void DoublePendulum::update() {
 }
 
 Vector2 DoublePendulum::getPos1(Vector2 origin) const {
-    return {origin.x + l1 * sin(a1), origin.y + l1 * cos(a1)};
+    return {(float)(origin.x + l1 * sin(a1)), (float)(origin.y + l1 * cos(a1))};
 }
 
 Vector2 DoublePendulum::getPos2(Vector2 origin) const {
     Vector2 pos1 = getPos1(origin);
-    return {pos1.x + l2 * sin(a2), pos1.y + l2 * cos(a2)};
+    return {(float)(pos1.x + l2 * sin(a2)), (float)(pos1.y + l2 * cos(a2))};
 }
